fix: fix custom constructors conflict
All constructors are functions that are attached to DataType. However,
when user defines a function, its `this` type is always Any, so that
method dispatch cannot select the proper function when type is called.

The example below would call Y’s constructor (Y(0)) instead of X’s
because it was defined last.
```
type X = {x}
type Y = {y}

fn X() = X(0)
fn Y() = Y(0)

println(X())
```

Fix this by inspecting the type of the function and using either Any
or Type{T} as a type of `this` argument.

(This was a known regression from HIR refactoring.)